test(syntax): merge module scan tests, marker-based multi-file fixtures
