+++ tests/unit/syntax/module_scan_tests.cpp
@@ -1,12 +1,59 @@
-#include "module_scan_fixture.h"
+#include <cassert>
+
 #include "test/test.h"
+#include "syntax/annotation.h"
 #include "syntax/scan.h"
 
 namespace clice::testing {
 namespace {
 
+/// Sets up a TestVFS from one `#[file]`-marked source blob (the unit-test
+/// multi-file grammar, split by AnnotatedSources::add_sources) and calls
+/// the given scan function with standard C++20 arguments. The first
+/// marked file is the scanned main file.
+struct ModuleScanFixture {
+    llvm::IntrusiveRefCntPtr<TestVFS> vfs = llvm::makeIntrusiveRefCnt<TestVFS>();
+    std::string main_path;
+    std::vector<const char*> args;
+
+    // Non-copyable/non-movable: args stores raw pointers (main_path.c_str())
+    // that would dangle after copy/move.
+    ModuleScanFixture(const ModuleScanFixture&) = delete;
+    ModuleScanFixture& operator=(const ModuleScanFixture&) = delete;
+    ModuleScanFixture(ModuleScanFixture&&) = delete;
+    ModuleScanFixture& operator=(ModuleScanFixture&&) = delete;
+
+    ModuleScanFixture(llvm::StringRef content, std::initializer_list<const char*> extra_args = {}) {
+        auto begin = content.find("#[");
+        assert(begin != llvm::StringRef::npos && "fixture content needs a #[file] marker");
+        auto end = content.find(']', begin);
+        main_path = TestVFS::path(content.slice(begin + 2, end));
+
+        AnnotatedSources sources;
+        sources.add_sources(content);
+        for(const auto& entry: sources.all_files) {
+            vfs->add(entry.getKey(), entry.getValue().content);
+        }
+
+        args.push_back("clang++");
+        args.push_back("-std=c++20");
+        for(auto arg: extra_args) {
+            args.push_back(arg);
+        }
+        args.push_back(main_path.c_str());
+    }
+
+    ScanResult decl() {
+        return scan_module_decl(args, TestVFS::root(), {}, nullptr, vfs);
+    }
+
+    ScanResult precise() {
+        return scan_precise(args, TestVFS::root(), {}, nullptr, vfs);
+    }
+};
+
 // =============================================================================
-// scan() — module declaration extraction (lexer-based, cppref coverage)
+// scan_quick() — module declaration extraction (lexer-based)
 // =============================================================================
 
 TEST_SUITE(ModuleScan) {
@@ -130,6 +177,21 @@ export module mylib;
     EXPECT_FALSE(result.need_preprocess);
 }
 
+// Imports are deliberately NOT collected by the quick scan: unlike the
+// module declaration name (which [cpp.module]/2 forbids to be an
+// object-like macro), an import's tokens are macro-expanded, so
+// directive-level text is not a trustworthy source of dependency edges —
+// scan_precise() is (see ImportMacroExpandedName below).
+TEST_CASE(ImportsNotCollected) {
+    auto result = scan_quick(R"(
+export module top;
+import base;
+export import my.nested:part;
+)");
+    EXPECT_EQ(result.module_name, "top");
+    EXPECT_TRUE(result.modules.empty());
+}
+
 // Private module fragment marker should not override the real module declaration.
 TEST_CASE(PrivateModuleFragment) {
     auto result = scan_quick(R"(
@@ -151,7 +213,10 @@ int f() { return 42; }
 TEST_SUITE(ModuleDeclFallback) {
 
 TEST_CASE(Basic) {
-    ModuleScanFixture f("main.cppm", "export module mylib;");
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+)");
     auto result = f.decl();
     EXPECT_EQ(result.module_name, "mylib");
     EXPECT_TRUE(result.is_interface_unit);
@@ -159,7 +224,8 @@ TEST_CASE(Basic) {
 
 TEST_CASE(ConditionalWithDefine) {
     // Without -DUSE_MODULES: no module declaration.
-    ModuleScanFixture f1("main.cppm", R"(
+    ModuleScanFixture f1(R"(
+#[main.cppm]
 #ifdef USE_MODULES
 export module mylib;
 #endif
@@ -167,8 +233,8 @@ export module mylib;
     EXPECT_TRUE(f1.decl().module_name.empty());
 
     // With -DUSE_MODULES: module declaration found.
-    ModuleScanFixture f2("main.cppm",
-                         R"(
+    ModuleScanFixture f2(R"(
+#[main.cppm]
 #ifdef USE_MODULES
 export module mylib;
 #endif
@@ -181,7 +247,8 @@ export module mylib;
 
 TEST_CASE(ConditionalIfExpr) {
     // Without the define: no module.
-    ModuleScanFixture f1("main.cppm", R"(
+    ModuleScanFixture f1(R"(
+#[main.cppm]
 #if ENABLE_MODULES >= 1
 export module mylib;
 #endif
@@ -189,8 +256,8 @@ export module mylib;
     EXPECT_TRUE(f1.decl().module_name.empty());
 
     // With the define: module found.
-    ModuleScanFixture f2("main.cppm",
-                         R"(
+    ModuleScanFixture f2(R"(
+#[main.cppm]
 #if ENABLE_MODULES >= 1
 export module mylib;
 #endif
@@ -202,42 +269,56 @@ export module mylib;
 }
 
 TEST_CASE(GMFWithConditional) {
-    ModuleScanFixture f("main.cppm", R"(
+    ModuleScanFixture f(R"(
+#[main.cppm]
 module;
 #include "config.h"
 #ifdef USE_MODULES
 export module mylib;
 #endif
+#[config.h]
+#define USE_MODULES 1
 )");
-    f.add_file("config.h", "#define USE_MODULES 1\n");
     auto result = f.decl();
     EXPECT_EQ(result.module_name, "mylib");
     EXPECT_TRUE(result.is_interface_unit);
 }
 
 TEST_CASE(ImplementationUnit) {
-    ModuleScanFixture f("main.cpp", "module mylib;");
+    ModuleScanFixture f(R"(
+#[main.cpp]
+module mylib;
+)");
     auto result = f.decl();
     EXPECT_EQ(result.module_name, "mylib");
     EXPECT_FALSE(result.is_interface_unit);
 }
 
 TEST_CASE(DottedName) {
-    ModuleScanFixture f("main.cppm", "export module std.io;");
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module std.io;
+)");
     auto result = f.decl();
     EXPECT_EQ(result.module_name, "std.io");
     EXPECT_TRUE(result.is_interface_unit);
 }
 
 TEST_CASE(Partition) {
-    ModuleScanFixture f("main.cppm", "export module mylib:core;");
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib:core;
+)");
     auto result = f.decl();
     EXPECT_EQ(result.module_name, "mylib:core");
     EXPECT_TRUE(result.is_interface_unit);
 }
 
 TEST_CASE(NoModule) {
-    ModuleScanFixture f("main.cpp", "int main() { return 0; }");
+    ModuleScanFixture f(R"(
+#[main.cpp]
+int main() { return 0; }
+)");
     auto result = f.decl();
     EXPECT_TRUE(result.module_name.empty());
     EXPECT_FALSE(result.is_interface_unit);
@@ -245,6 +326,331 @@ TEST_CASE(NoModule) {
 }
 
 };  // TEST_SUITE(ModuleDeclFallback)
+
+// =============================================================================
+// scan_precise() — module import semantics
+// =============================================================================
+
+TEST_SUITE(ModuleImportScan) {
+
+TEST_CASE(NamedImport) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+import other;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    EXPECT_TRUE(result.is_interface_unit);
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "other");
+}
+
+TEST_CASE(MultipleImports) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+import alpha;
+import beta;
+import gamma;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_EQ(result.modules.size(), 3u);
+    EXPECT_EQ(result.modules[0], "alpha");
+    EXPECT_EQ(result.modules[1], "beta");
+    EXPECT_EQ(result.modules[2], "gamma");
+}
+
+TEST_CASE(DottedModuleImport) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+import std.io;
+)");
+    auto result = f.precise();
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "std.io");
+}
+
+// Partition import: clang returns the fully-qualified name "mylib:core"
+// (owning module + ':' + partition name) as a single ModuleIdPath entry.
+TEST_CASE(PartitionImport) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+import :core;
+)");
+    auto result = f.precise();
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "mylib:core");
+}
+
+// Export-import of a named module.
+TEST_CASE(ExportImport) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+export import other;
+)");
+    auto result = f.precise();
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "other");
+}
+
+// Export-import of a partition.
+TEST_CASE(ExportImportPartition) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+export import :core;
+)");
+    auto result = f.precise();
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "mylib:core");
+}
+
+// Implementation unit importing a named module.
+TEST_CASE(ImplementationImport) {
+    ModuleScanFixture f(R"(
+#[impl.cpp]
+module mylib;
+import other;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    EXPECT_FALSE(result.is_interface_unit);
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "other");
+}
+
+// Implementation unit importing a partition of the same module.
+TEST_CASE(ImplementationPartitionImport) {
+    ModuleScanFixture f(R"(
+#[impl.cpp]
+module mylib;
+import :utils;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "mylib:utils");
+}
+
+// Multiple partition imports.
+TEST_CASE(MultiplePartitionImports) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+export import :core;
+import :utils;
+import :io;
+)");
+    auto result = f.precise();
+    ASSERT_EQ(result.modules.size(), 3u);
+    EXPECT_EQ(result.modules[0], "mylib:core");
+    EXPECT_EQ(result.modules[1], "mylib:utils");
+    EXPECT_EQ(result.modules[2], "mylib:io");
+}
+
+// Mixed named module imports and partition imports.
+TEST_CASE(MixedNamedAndPartitionImports) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+import other;
+export import :core;
+import another.lib;
+import :utils;
+)");
+    auto result = f.precise();
+    ASSERT_EQ(result.modules.size(), 4u);
+    EXPECT_EQ(result.modules[0], "other");
+    EXPECT_EQ(result.modules[1], "mylib:core");
+    EXPECT_EQ(result.modules[2], "another.lib");
+    EXPECT_EQ(result.modules[3], "mylib:utils");
+}
+
+// NOTE: Header unit imports (import <header>; / import "header";) are not
+// tested here because they require actual header unit compilation support
+// which clang's PreprocessOnlyAction doesn't provide in a VFS-only context.
+// These would hang trying to resolve system headers.
+
+// GMF with imports.
+TEST_CASE(GMFWithImport) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+module;
+#include "config.h"
+export module mylib;
+import dep;
+#[config.h]
+// config
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    EXPECT_TRUE(result.is_interface_unit);
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "dep");
+}
+
+// Mixed includes (from GMF) and imports (after module decl).
+TEST_CASE(MixedIncludesAndImports) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+module;
+#include "legacy.h"
+export module mylib;
+import dep_a;
+import dep_b;
+export int f();
+#[legacy.h]
+int legacy_func();
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_GE(result.includes.size(), 1u);
+    ASSERT_EQ(result.modules.size(), 2u);
+    EXPECT_EQ(result.modules[0], "dep_a");
+    EXPECT_EQ(result.modules[1], "dep_b");
+}
+
+// No module — plain C++ file.
+TEST_CASE(NoModule) {
+    ModuleScanFixture f(R"(
+#[main.cpp]
+#include "header.h"
+int main() { return 0; }
+#[header.h]
+int x;
+)");
+    auto result = f.precise();
+    EXPECT_TRUE(result.module_name.empty());
+    EXPECT_FALSE(result.is_interface_unit);
+    EXPECT_TRUE(result.modules.empty());
+}
+
+// Partition interface unit declaring and importing another partition.
+TEST_CASE(PartitionInterfaceImportingPartition) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib:ui;
+import :core;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib:ui");
+    EXPECT_TRUE(result.is_interface_unit);
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "mylib:core");
+}
+
+// Partition implementation importing another partition.
+TEST_CASE(PartitionImplImportingPartition) {
+    ModuleScanFixture f(R"(
+#[impl.cpp]
+module mylib:detail;
+import :core;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib:detail");
+    EXPECT_FALSE(result.is_interface_unit);
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "mylib:core");
+}
+
+// Conditional imports: the precise scan evaluates the condition, so only
+// the taken branch's import is a dependency edge.
+TEST_CASE(ConditionalImport) {
+    ModuleScanFixture f1(R"(
+#[main.cppm]
+export module mylib;
+#ifdef WITH_DEP
+import dep;
+#endif
+)");
+    EXPECT_TRUE(f1.precise().modules.empty());
+
+    ModuleScanFixture f2(R"(
+#[main.cppm]
+export module mylib;
+#ifdef WITH_DEP
+import dep;
+#endif
+)",
+                         {"-DWITH_DEP"});
+    auto result = f2.precise();
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "dep");
+}
+
+// Import target is a macro-expanded name.
+// C++20 forbids object-like macros in module DECLARATIONS (export module M;),
+// but clang's preprocessor expands macros in import declarations —
+// [cpp.import]: an import's tokens are "processed just as in normal text".
+TEST_CASE(ImportMacroExpandedName) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+#define OTHER_MOD other
+import OTHER_MOD;
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "other");
+}
+
+// Import target from a macro defined on the command line.
+TEST_CASE(ImportMacroFromCommandLine) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+import DEP_MOD;
+)",
+                        {"-DDEP_MOD=dependency"});
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "dependency");
+}
+
+// Import target from a macro defined in GMF header.
+TEST_CASE(ImportMacroFromGMFHeader) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+module;
+#include "deps.h"
+export module mylib;
+import MY_DEP;
+#[deps.h]
+#define MY_DEP some_lib
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "some_lib");
+}
+
+// Import target from a macro defined in a header #included AFTER the module
+// declaration (not in GMF). C++20 allows #include after module declarations —
+// the preprocessor still processes them and any macros they define are visible
+// to subsequent import declarations.
+TEST_CASE(ImportMacroFromPostDeclInclude) {
+    ModuleScanFixture f(R"(
+#[main.cppm]
+export module mylib;
+#include "imports.h"
+import MY_IMPORT;
+#[imports.h]
+#define MY_IMPORT dep
+)");
+    auto result = f.precise();
+    EXPECT_EQ(result.module_name, "mylib");
+    ASSERT_EQ(result.modules.size(), 1u);
+    EXPECT_EQ(result.modules[0], "dep");
+}
+
+};  // TEST_SUITE(ModuleImportScan)
 
 }  // namespace
 }  // namespace clice::testing
